Rename LinqToDB related packages and references. Rule from official maintainers of LinqToDB: `linq2db` for NuGet packages, everything else is `LinqToDB`.
