+++ src/ByteAether.Ulid.slnx
@@ -10,9 +10,9 @@
     <Project Path="EFCore.Tests/EFCore.Tests.csproj" />
     <Project Path="EFCore/EFCore.Impl.csproj" />
   </Folder>
-  <Folder Name="/Linq2Db/">
-    <Project Path="Linq2Db.IntegrationTests/Linq2Db.IntegrationTests.csproj" />
-    <Project Path="Linq2Db/Linq2Db.Impl.csproj" />
+  <Folder Name="/LinqToDB/">
+    <Project Path="LinqToDB\LinqToDB.Impl.csproj" />
+    <Project Path="LinqToDB.IntegrationTests\LinqToDB.IntegrationTests.csproj" />
   </Folder>
   <Folder Name="/Solution Items/">
     <File Path="../README.md" />
